fix (NSObject+PXSubclass) remove the NSAssert(NO ...) that crashes the app when subclassInstance is called with an class that is already extended by pixate
I have no idea why the assertion is there in the first place.
NSAssert(NO, ./.. will always crash.. is not a valid use for NSAssert..
Also there is a return after the assert .. maybe the assert was put there for debug purposes.. like: "this should almost never happen.. so when it happens it will crash and we will have a look into it " , but this is just speculative.. I have no ideas why it was there.